The Java language specification makes these type casts redundant

+++ src/main/java/com/fasterxml/jackson/core/Base64Variant.java
@@ -170,7 +170,7 @@ public final class Base64Variant
 
         // Plus if we use padding, add that in too
         if (writePadding) {
-            _asciiToBase64[(int) paddingChar] = BASE64_VALUE_PADDING;
+            _asciiToBase64[paddingChar] = BASE64_VALUE_PADDING;
         }
 
         // By default, require padding on input if written on output; do not
@@ -342,7 +342,7 @@ public final class Base64Variant
     }
 
     public boolean usesPaddingChar(char c) { return c == _paddingChar; }
-    public boolean usesPaddingChar(int ch) { return ch == (int) _paddingChar; }
+    public boolean usesPaddingChar(int ch) { return ch == _paddingChar; }
 
     /**
      * @return Indicator on how this Base64 encoding will handle possible padding
@@ -370,7 +370,7 @@ public final class Base64Variant
      */
     public int decodeBase64Char(char c)
     {
-        int ch = (int) c;
+        int ch = c;
         return (ch <= 127) ? _asciiToBase64[ch] : BASE64_VALUE_INVALID;
     }
 
@@ -381,7 +381,7 @@ public final class Base64Variant
 
     public int decodeBase64Byte(byte b)
     {
-        int ch = (int) b;
+        int ch = b;
         // note: cast retains sign, so it's from -128 to +127
         if (ch < 0) {
             return BASE64_VALUE_INVALID;
@@ -578,9 +578,9 @@ public final class Base64Variant
 
         while (inputPtr <= safeInputEnd) {
             // First, mash 3 bytes into lsb of 32-bit int
-            int b24 = ((int) input[inputPtr++]) << 8;
-            b24 |= ((int) input[inputPtr++]) & 0xFF;
-            b24 = (b24 << 8) | (((int) input[inputPtr++]) & 0xFF);
+            int b24 = (input[inputPtr++]) << 8;
+            b24 |= (input[inputPtr++]) & 0xFF;
+            b24 = (b24 << 8) | ((input[inputPtr++]) & 0xFF);
             encodeBase64Chunk(sb, b24);
             if (--chunksBeforeLF <= 0) {
                 // note: must quote in JSON value, so not really useful...
@@ -593,9 +593,9 @@ public final class Base64Variant
         // And then we may have 1 or 2 leftover bytes to encode
         int inputLeft = inputEnd - inputPtr; // 0, 1 or 2
         if (inputLeft > 0) { // yes, but do we have room for output?
-            int b24 = ((int) input[inputPtr++]) << 16;
+            int b24 = (input[inputPtr++]) << 16;
             if (inputLeft == 2) {
-                b24 |= (((int) input[inputPtr++]) & 0xFF) << 8;
+                b24 |= ((input[inputPtr++]) & 0xFF) << 8;
             }
             encodeBase64Partial(sb, b24, inputLeft);
         }
@@ -633,9 +633,9 @@ public final class Base64Variant
         int safeInputEnd = inputEnd-3;
 
         while (inputPtr <= safeInputEnd) {
-            int b24 = ((int) input[inputPtr++]) << 8;
-            b24 |= ((int) input[inputPtr++]) & 0xFF;
-            b24 = (b24 << 8) | (((int) input[inputPtr++]) & 0xFF);
+            int b24 = (input[inputPtr++]) << 8;
+            b24 |= (input[inputPtr++]) & 0xFF;
+            b24 = (b24 << 8) | ((input[inputPtr++]) & 0xFF);
             encodeBase64Chunk(sb, b24);
             if (--chunksBeforeLF <= 0) {
                 sb.append(linefeed);
@@ -644,9 +644,9 @@ public final class Base64Variant
         }
         int inputLeft = inputEnd - inputPtr;
         if (inputLeft > 0) {
-            int b24 = ((int) input[inputPtr++]) << 16;
+            int b24 = (input[inputPtr++]) << 16;
             if (inputLeft == 2) {
-                b24 |= (((int) input[inputPtr++]) & 0xFF) << 8;
+                b24 |= ((input[inputPtr++]) & 0xFF) << 8;
             }
             encodeBase64Partial(sb, b24, inputLeft);
         }
