+++ src/main/java/com/fasterxml/jackson/core/json/UTF8JsonGenerator.java
@@ -771,7 +771,7 @@ public class UTF8JsonGenerator
         while (offset < len) {
             inner_loop:
             while (true) {
-                int ch = (int) cbuf[offset];
+                int ch = cbuf[offset];
                 if (ch > 0x7F) {
                     break inner_loop;
                 }
@@ -821,7 +821,7 @@ public class UTF8JsonGenerator
         while (offset < inputEnd) {
             inner_loop:
             while (true) {
-                int ch = (int) cbuf[offset];
+                int ch = cbuf[offset];
                 if (ch >= 0x80) {
                     break inner_loop;
                 }
@@ -862,7 +862,7 @@ public class UTF8JsonGenerator
         while (offset < end) {
             inner_loop:
             while (true) {
-                int ch = (int) cbuf[offset];
+                int ch = cbuf[offset];
                 if (ch > 0x7F) {
                     break inner_loop;
                 }
@@ -1916,9 +1916,9 @@ public class UTF8JsonGenerator
                 _flushBuffer();
             }
             // First, mash 3 bytes into lsb of 32-bit int
-            int b24 = ((int) input[inputPtr++]) << 8;
-            b24 |= ((int) input[inputPtr++]) & 0xFF;
-            b24 = (b24 << 8) | (((int) input[inputPtr++]) & 0xFF);
+            int b24 = (input[inputPtr++]) << 8;
+            b24 |= (input[inputPtr++]) & 0xFF;
+            b24 = (b24 << 8) | ((input[inputPtr++]) & 0xFF);
             _outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer, _outputTail);
             if (--chunksBeforeLF <= 0) {
                 // note: must quote in JSON value
@@ -1934,9 +1934,9 @@ public class UTF8JsonGenerator
             if (_outputTail > safeOutputEnd) { // don't really need 6 bytes but...
                 _flushBuffer();
             }
-            int b24 = ((int) input[inputPtr++]) << 16;
+            int b24 = (input[inputPtr++]) << 16;
             if (inputLeft == 2) {
-                b24 |= (((int) input[inputPtr++]) & 0xFF) << 8;
+                b24 |= ((input[inputPtr++]) & 0xFF) << 8;
             }
             _outputTail = b64variant.encodeBase64Partial(b24, inputLeft, _outputBuffer, _outputTail);
         }
@@ -1967,9 +1967,9 @@ public class UTF8JsonGenerator
             if (_outputTail > safeOutputEnd) { // need to flush
                 _flushBuffer();
             }
-            int b24 = ((int) readBuffer[inputPtr++]) << 8;
-            b24 |= ((int) readBuffer[inputPtr++]) & 0xFF;
-            b24 = (b24 << 8) | (((int) readBuffer[inputPtr++]) & 0xFF);
+            int b24 = (readBuffer[inputPtr++]) << 8;
+            b24 |= (readBuffer[inputPtr++]) & 0xFF;
+            b24 = (b24 << 8) | ((readBuffer[inputPtr++]) & 0xFF);
             bytesLeft -= 3;
             _outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer, _outputTail);
             if (--chunksBeforeLF <= 0) {
@@ -1987,10 +1987,10 @@ public class UTF8JsonGenerator
                 if (_outputTail > safeOutputEnd) { // don't really need 6 bytes but...
                     _flushBuffer();
                 }
-                int b24 = ((int) readBuffer[inputPtr++]) << 16;
+                int b24 = (readBuffer[inputPtr++]) << 16;
                 int amount;
                 if (inputPtr < inputEnd) {
-                    b24 |= (((int) readBuffer[inputPtr]) & 0xFF) << 8;
+                    b24 |= ((readBuffer[inputPtr]) & 0xFF) << 8;
                     amount = 2;
                 } else {
                     amount = 1;
@@ -2030,9 +2030,9 @@ public class UTF8JsonGenerator
                 _flushBuffer();
             }
             // First, mash 3 bytes into lsb of 32-bit int
-            int b24 = ((int) readBuffer[inputPtr++]) << 8;
-            b24 |= ((int) readBuffer[inputPtr++]) & 0xFF;
-            b24 = (b24 << 8) | (((int) readBuffer[inputPtr++]) & 0xFF);
+            int b24 = (readBuffer[inputPtr++]) << 8;
+            b24 |= (readBuffer[inputPtr++]) & 0xFF;
+            b24 = (b24 << 8) | ((readBuffer[inputPtr++]) & 0xFF);
             bytesDone += 3;
             _outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer, _outputTail);
             if (--chunksBeforeLF <= 0) {
@@ -2047,10 +2047,10 @@ public class UTF8JsonGenerator
             if (_outputTail > safeOutputEnd) { // don't really need 6 bytes but...
                 _flushBuffer();
             }
-            int b24 = ((int) readBuffer[inputPtr++]) << 16;
+            int b24 = (readBuffer[inputPtr++]) << 16;
             int amount = 1;
             if (inputPtr < inputEnd) {
-                b24 |= (((int) readBuffer[inputPtr]) & 0xFF) << 8;
+                b24 |= ((readBuffer[inputPtr]) & 0xFF) << 8;
                 amount = 2;
             }
             bytesDone += amount;
